+++ sdk/core/Microsoft.Azure.Core.Spatial.NewtonsoftJson/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ### Added
 
-- Added `NewtonsoftJsonGeographyPointConverter` to serialize `Microsoft.Spatial.GeographyPoint` objects.
+- Added `NewtonsoftJsonGeographyConverter` to serialize `Microsoft.Spatial.GeographyPoint` objects.